try to remove __int64 which is only in MSVC

+++ src/DumpUtil.h
@@ -1005,7 +1005,7 @@ SOFTWARE.
 
 #define NAV_FIELD_PROP_NUMBER64_DESC_F(Field_Name, Field_Bits, Field_Value, Field_Desc, ...)\
 	MBCSPRINTF_S(szTemp2, TEMP2_SIZE, Field_Desc, ##__VA_ARGS__);\
-	MBCSPRINTF_S(szTemp3, TEMP3_SIZE, "%llu", (__int64)(Field_Value));\
+	MBCSPRINTF_S(szTemp3, TEMP3_SIZE, "%lld", (long long)(Field_Value));\
 	NAV_FIELD_PROP(Field_Name, Field_Bits, szTemp3, szTemp2, bit_offset?*bit_offset:-1LL, "I");\
 	if (bit_offset)*bit_offset += Field_Bits;\
 
@@ -1090,7 +1090,7 @@ SOFTWARE.
 	if (bit_offset)*bit_offset += Field_Bits;\
 
 #define NAV_FIELD_PROP_NUMBER64(Field_Name, Field_Bits, Field_Value, Field_Desc)\
-	MBCSPRINTF_S(szTemp3, TEMP3_SIZE, "%llu", (__int64)(Field_Value));\
+	MBCSPRINTF_S(szTemp3, TEMP3_SIZE, "%lld", (long long)(Field_Value));\
 	NAV_FIELD_PROP(Field_Name, Field_Bits, szTemp3, Field_Desc, bit_offset?*bit_offset:-1LL, "I");\
 	if (bit_offset)*bit_offset += Field_Bits;\
